override_dh_install: Use make conditionals rather than shell
Either make syntax or shell syntax is fine, but the use of $(BUILD_DOCS)
needs to be consistent with that in override_dh_auto_build-indep.  In
particular, make's $(if ...) considers '0' to be true (empty string is
false).

This change switches to make syntax (empty string is false), which is a
bit more surprising but has the advantage that we can drop all the "; \"
nonsense at end of line (and gain `set -e` behaviour).