Rebuild hoedown html_blocks.c using gperf. Don't exclude jquery for now.
We need to think through how to do jquery in more depth. We can't just have
rustdoc copy the system jquery instead, since the rust-doc build uses that, and
we will still end up with duplicate copies in Debian. But having rustdoc
instead create a symlink is a more invasive change from upstream.